+++ src/components/Title.tsx
@@ -1,0 +1,10 @@
+const Title = ()=> (
+    <h1 style={{
+        textAlign: 'center',
+        textShadow:'1px 1px 1px black'
+    }}>バトルゲーム
+    </h1>
+);
+
+export default Title;
+
